[FlexNG] Fix great-grandchild relative-position offsets.
In order to determine the correct size to resolve %-based relative
top offsets against we need to determine if the available-size given
to a flex-item is definite.

Previously this worked as there was a bit on the LayoutBox which
captured this information. This patch for NG flex-items allows for
looking at the previous constraint space while outside of layout.

Bug: 1108508
Change-Id: I3104b7afd4d96b707a8110eb2ed5d59d17129580